fix: Windows MSVC compatibility for xhttp tests
Key fixes:
- xIOBlock.refs: size_t -> long to match MSVC 32-bit Interlocked* ops
- Block pool & logger freelist: xAtomicCasWeak -> xAtomicCasPtrWeak
  for correct 64-bit pointer CAS on Windows
- Add NOMINMAX, xnet_socket_t, XNET_INVALID_SOCKET, xnet_strndup
  to xnet/compat.h
- Add XK_EXPORT on transport private functions for DLL export
- Use xnet_strndup in server.c, sse.c, proto_h2.c
- Use xnet_socket_t in test helpers
- Conditionally disable -Wno-error for MSVC in cmake modules
- Fix <windows.h> include order in backtrace.c

Test results: 134/156 pass (22 failures are TLS/network env only)

+++ modules/xbuf/io.h
@@ -68,7 +68,7 @@
  * Users should not manipulate blocks directly.
  */
 XDEF_STRUCT(xIOBlock) {
-  size_t refs;                       /**< Reference count (atomic)     */
+  long   refs;                       /**< Reference count (atomic)     */
   size_t size;                       /**< Usable data size in bytes    */
   char   data[XIOBUFFER_BLOCK_SIZE]; /**< Inline data storage          */
 };
